let's instead just keep track of the oldest synitem date in the syndicated table so we can use it later to do a mass cleaning, or part of a migration to a clustered synitem table.
those deletes are horrendous on a heavily fragmented myisam table.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4779 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ bin/maint/synsuck.pl
@@ -346,9 +346,13 @@ $maint{'synsuck'} = sub
         # entry, it's possible for the oldest item that was previously
         # gone to reappear, and we want to protect against that a
         # little.
-        $dbh->do("DELETE FROM synitem WHERE userid=? AND ".
-                 "dateadd < ? - INTERVAL 14 DAY LIMIT 1000",
-                 undef, $userid, $mindate);
+        if ($LJ::SYNITEM_CLEAN) {
+            $dbh->do("DELETE FROM synitem WHERE userid=? AND ".
+                     "dateadd < ? - INTERVAL 14 DAY LIMIT 1000",
+                     undef, $userid, $mindate);
+        }
+        $dbh->do("UPDATE syndicated SET oldest_ourdate=? WHERE userid=?",
+                 undef, $mindate, $userid);
 
         # bail out if errors, and try again shortly
         if ($errorflag) {
